infra/requirements.txt
aws-cdk-lib==2.148.0
constructs==10.3.0
